third_party/WebKit/LayoutTests/media/media-basic-suspend.html

<!DOCTYPE html>
<title>Verify that a media element can be resumed after idle suspend.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-file.js"></script>
<video></video>
<script>
async_test(function(t) {
  var video = document.querySelector('video');
  var mediaFile = findMediaFile('video', 'content/test');

  video.ontimeupdate = video.onended = t.step_func(function() {
    if (video.currentTime > 0) {
      assert_false(window.internals.isMediaElementSuspended(video));
      t.done();
    }
  });

  video.oncanplaythrough = t.step_func(function() {
    assert_true(!!window.internals,
                'This test requires internals support for suspend/resume.');

    video.addEventListener('suspended', t.step_func(function() {
      assert_true(window.internals.isMediaElementSuspended(video));
      video.play();
    }, false));

    window.internals.forceStaleStateForMediaElement(video);
  });

  video.src = mediaFile;
});
</script>
